cleaned up example again
